requirements.txt
Django==3.1.7
djangorestframework==3.12.4
django-cors-headers==3.2.1
pytz==2019.3
sqlparse==0.3.0
django-rest-knox==4.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==1.7.1
graphene~=2.1.8
django-filter==2.4.0
graphene-sqlalchemy==2.3.0
django-graphql-jwt==0.3.0
django-redis==4.12.1
django-graphql-auth==0.3.16
graphene-django==2.15.0
pytest==3.10.0
pytest-django==3.1.2